\name{DSC2014Tutorial-package}
\alias{DSC2014Tutorial-package}
\docType{package}
\title{
Materials of Tutorial Program on DSC 2014
}
\description{
This package contains the required materials of R Tutorial DSC2014
}
\details{
\tabular{ll}{
Package: \tab DSC2014Tutorial\cr
Type: \tab Package\cr
Version: \tab 0.5.1\cr
Date: \tab 2014-08-03\cr
License: \tab GPL (>= 3)\cr
}

}
\author{
Taiwan R User Group\cr
Wush Wu <wush978@gmail.com>
}
