\name{slides}
\alias{slides}
\title{Demonstrate the course slide}
\usage{
  slides(x)
}
\arguments{
  \item{x}{a character of course name.}
}
\value{
  demonstrate a HTML5 slide in your default browser
}
\description{
  \code{slides} demonstrate the HTML5 slide for the selected course.
}
\examples{

slides("Basic") # For R Basic
slides("ETL1") # For ETL on R - Step 1 and 2
slides("ETL2") # For ETL on R - step 3 and 4
slides("DataAnalysis") # For DataA nalysis  with R
slides("Visualization1") # For R Basic Visualization - Easy Way to Plot
slides("Visualization2") # For Introductory R Visualization - Using ggplot2
slides("Visualization3") # For An Introduction to R Markdown
}

